libpmi: move from libpmi-server
Begin consolidating PMI code in src/common/libpmi.
This first commit moves src/common/libpmi-server into libpmi,
and builds a libpmi-simple.la internal library.

Update users.
